# subscription-service/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Subscription Service API
  version: 1.0.0
paths:
  /:
    get:
      summary: Check service status
      responses:
        "200":
          description: Service is running

  /subscriptions:
    post:
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscription'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        "400":
          description: Invalid request
        "500":
          description: Internal server error

    get:
      summary: List subscriptions
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
        - in: query
          name: service_name
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        "500":
          description: Internal server error

  /subscriptions/{id}:
    get:
      summary: Get subscription by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        "404":
          description: Not found
        "500":
          description: Internal server error

    put:
      summary: Update subscription by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscription'
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        "400":
          description: Invalid request
        "404":
          description: Not found
        "500":
          description: Internal server error

    delete:
      summary: Delete subscription by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content (deleted)
        "404":
          description: Not found
        "500":
          description: Internal server error

  /subscriptions/aggregate:
    get:
      summary: Get total subscription cost for period
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            example: "07-2025"
        - in: query
          name: to
          required: true
          schema:
            type: string
            example: "09-2025"
        - in: query
          name: user_id
          schema:
            type: string
        - in: query
          name: service_name
          schema:
            type: string
      responses:
        "200":
          description: Total cost
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:
                    type: string
                  to:
                    type: string
                  total:
                    type: integer
        "400":
          description: Invalid request
        "500":
          description: Internal server error

components:
  schemas:
    CreateSubscription:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          nullable: true
          example: "12-2025"

    UpdateSubscription:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          nullable: true
          example: "12-2025"

    Subscription:
      type: object
      properties:
        id:
          type: string
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time